chore(comments): add minor enhancements
- add unique key attribute to tag list
- add getDisplayDate function to the helper functions for reusability
