+++ src/utils/helpers.js
@@ -1,5 +1,6 @@
 import jwt from 'jsonwebtoken';
-
+import simpleDateFormat from 'date-fns/format';
+import compareDateFormat from 'date-fns/distance_in_words_strict';
 /**
  * Extract a given url query value
  *
@@ -18,5 +19,23 @@ export const getUrlQuery = (param) => {
  * @returns {object} The decoded payload
  */
 export const verifyToken = token => jwt.verify(token, process.env.APP_KEY);
+
+/**
+   * Formats the date to a human readable format
+   * @param {object} date The date object to format
+   * @returns {object} An object  containing a short and long format representations of the date
+   */
+export const getDisplayDate = (date) => {
+  const today = new Date();
+  const result = {
+    short: simpleDateFormat(date, 'MMM Do, YYYY'),
+    long: simpleDateFormat(date, 'MMM Do, YYYY h:mm a')
+  };
+  if (simpleDateFormat(date, 'M YYYY') === simpleDateFormat(today, 'M YYYY')) {
+    const difference = compareDateFormat(date, today);
+    result.short = difference.startsWith('0') ? '1 second ago' : `${difference} ago`;
+  }
+  return result;
+};
 
 export default {};
